@@ -1,6 +1,6 @@
 module github.com/kak-tus/irma_bot
 
-go 1.17
+go 1.21
 
 require (
 	github.com/deepmap/oapi-codegen v1.11.0
